implementer les commentaires

+++ app/Http/Controllers/ArticleController.php
@@ -34,18 +34,10 @@ class ArticleController extends Controller
     {
         $validated = $this->validate($request, [
             'title' =>'required',
-            'content' =>'required',
-            'author' =>'required',
+            'content' =>'required'
         ]);
 
-        // $article = new Article();
-        // $article->title = $request->title;
-        // $article->content = $request->content;
-        // $article->author = $request->author;
-        // $article->save();
-
-        Article::create($validated);
-        //Article::create($request->all());
+        $request->user()->articles()->create($validated);
 
         return redirect()->route('articles.index')->with('success','article created successfully');
     }
@@ -63,6 +55,10 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        if(auth()->user()->id !== $article->user_id){
+            return redirect()->route('articles.index')->with('error','you are not authorized to edit this article');
+        }
+
         return view('articles.edit', compact('article'));
     }
 
@@ -71,17 +67,14 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        if($request->user()->id !== $article->user_id){
+            return redirect()->route('articles.index')->with('error','you are not authorized to edit this article');
+        }
  
         $validated = $this->validate($request, [
             'title' =>'required',
-            'content' =>'required',
-            'author' =>'required',
+            'content' =>'required'
         ]);
-
-        // $article->title = $request->title;
-        // $article->content = $request->content;
-        // $article->author = $request->author;
-        // $article->save();
 
         $article->update($validated);
 
@@ -93,6 +86,10 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        if(auth()->user()->id !== $article->user_id){
+            return redirect()->route('articles.index')->with('error','you are not authorized to edit this article');
+        }
+
         $article->delete();
 
         return redirect()->route('articles.index')->with('success','article deleted successfully');
